% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centerDS.R
\name{centerDS}
\alias{centerDS}
\title{Centering of a data frame}
\usage{
centerDS(table, ...)
}
\arguments{
\item{table}{\code{data.frame} or \code{numeric} vector to be centered}

\item{...}{\code{numeric} The means of the columns (ordered by column)}
}
\value{
\code{data.frame} or \code{numeric} vector centered
}
\description{
Centers each column of a data frame object by substracting the mean of the column
}
