% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummies_transformationDS.R
\name{dummies_transformationDS}
\alias{dummies_transformationDS}
\title{Transform table}
\usage{
dummies_transformationDS(x, ...)
}
\arguments{
\item{x}{\code{data.frame}}

\item{...}{\code{numerical} Vector to divide the table}
}
\value{
\code{data.table}
}
\description{
Divide columns of table by a vector of values,
this function is to be used to transform tables that only contain dummies for the FAMD functionality.
Other uses may yield to errors as there is no checks for coherence between the table and the
vector.
}
