% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummiesDS.R
\name{dummiesDS}
\alias{dummiesDS}
\title{Create dummy variables}
\usage{
dummiesDS(x)
}
\arguments{
\item{x}{\code{data.frame} Data frame to which create dummies for categorical variables}
}
\value{
\code{data.frame} Data frame with categorical variables transformed to dummies
}
\description{
Create dummy variables
}
