% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans.assign_resultDS.R
\name{kmeans.assign_resultDS}
\alias{kmeans.assign_resultDS}
\title{Assign values as a column on a data frame}
\usage{
kmeans.assign_resultDS(x, ...)
}
\arguments{
\item{x}{\code{data frame} Table where the next arguments will be added as a column}

\item{...}{\code{numeric} Values to be coerced as a column to the table 'x'}
}
\value{
\code{data frame} With the original 'x' table with a column added named 'kmeans.cluster'.
The added column is encoded as a factor and as such passes through the disclosive check: checking if the
number of levels of the factor is more than nfilter.levels (percentage) of the length of the variable.
}
\description{
This function is called from the ds.kmeans function when \code{assign = TRUE},
it could be adapted to other functionalities checking if lengths are equal.
}
