% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeansDS.R
\name{kmeansDS}
\alias{kmeansDS}
\title{Parallel k-means iteration}
\usage{
kmeansDS(x, ...)
}
\arguments{
\item{x}{\code{data frame} Train dataset for the k-means}

\item{...}{\code{numeric} Parameters corresponding to the data frame from the server that contains the
centroids (updated on each iteration on the client)}
}
\value{
\code{list} with: \cr
-counts \code{numeric} vector with the counts per cluster \cr
-centers \code{data frame} New centroids calculated \cr
-assignations \code{numeric vector} ordered cluster assignations, to be used by the client
to assign them with the \code{kmeans.assign_result} function to a table on the servers to be later used
}
\description{
Performs an iteration of a k-means parallel algorithm (what in a multi-thread machine
would be run on each thread). The client acts as the master and the servers as the slaves if thinking
like a regular parallel implementation.
}
