% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_dataset_treeDS.R
\name{prepare_dataset_treeDS}
\alias{prepare_dataset_treeDS}
\title{Title}
\usage{
prepare_dataset_treeDS(object, questions)
}
\arguments{
\item{object}{}

\item{questions}{}
}
\value{

}
\description{
Title
}
