% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaleDS.R
\name{scaleDS}
\alias{scaleDS}
\title{Scaling of a data frame or numeric vector}
\usage{
scaleDS(table, ...)
}
\arguments{
\item{table}{\code{data.frame} or \code{numeric} vector to be scaled}

\item{...}{\code{numeric} The standard deviations of the columns (ordered by column)}
}
\value{
\code{data.frame} or \code{numeric} vector scaled
}
\description{
Scales each column of a data frame object by dividing for the standard deviation of the column
}
