% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_typeDS.R
\name{subset_typeDS}
\alias{subset_typeDS}
\title{Subset data frame by class}
\usage{
subset_typeDS(table, type)
}
\arguments{
\item{table}{\code{data frame} Table to be subsetted}

\item{type}{\code{character} Class to which be subsetted. If \code{numeric} only the \code{numeric} columns
will be on the output table}
}
\value{
\code{data frame} Subsetted table
}
\description{
Create a subset of a data frame with the columns that have a shared class
}
