% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svdDS.R
\name{svdDS}
\alias{svdDS}
\title{Perform block operation SVD}
\usage{
svdDS(x)
}
\arguments{
\item{x}{\code{data frame} Data frame to perform the SVD}
}
\value{
\code{data frame} with the block SVD (left singluar vectors * singular values)
}
\description{
Perform block operation SVD
}
\details{
http://dx.doi.org/10.1137/16M1058467 For information on the SVD block method
}
