% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_dataset_treeDS.R
\name{prepare_dataset_treeDS}
\alias{prepare_dataset_treeDS}
\title{Title}
\usage{
prepare_dataset_treeDS(object, questions)
}
\arguments{
\item{object}{basis of dataset tree}

\item{questions}{criterion for preperation}
}
\value{
the prepared dataset tree
}
\description{
Title
}
